На линуксе unsigned long это нечто другое...

+++ builder/test-progs/multi/module-1.cpp
@@ -1,5 +1,5 @@
 #include "module-1.hpp"
 
-unsigned long func_1(unsigned long in) {
+u32 func_1(u32 in) {
   return in * 30;
 }
