+++ builder/test-progs/multi/module-1.hpp
@@ -1,3 +1,4 @@
 #pragma once
+#include "types.hpp"
 
-unsigned long func_1(unsigned long in);
+u32 func_1(u32 in);
